requirements.txt
requests
flask
pyyaml
python-dotenv
